requirements.txt
absl-py==2.3.0
anyio==3.7.1
astunparse==1.6.3
audioread==3.0.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==7.1.2
decorator==5.2.1
Distance==0.1.3
exceptiongroup==1.3.0
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==1.12
forcealign==1.1.9
fsspec==2025.5.1
future==1.0.0
g2p-en==2.1.0
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.72.1
h11==0.12.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.1.0
idna==3.10
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.1
keras==2.9.0
Keras-Preprocessing==1.1.2
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
lyricsgenius==3.6.4
Markdown==3.8
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
nltk==3.9.1
norbert==0.2.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==25.0
pandas==1.5.3
platformdirs==4.3.8
pooch==1.8.2
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
sox==1.5.0
soxr==0.5.0.post1
spleeter==2.4.0
sympy==1.14.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
triton==3.3.1
typeguard==4.4.3
typer==0.3.2
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
yt-dlp==2025.5.22
